types: change type of "block_height" from String to u128

+++ source/sak_p2p_disc/src/v0/ops/whoareyou/whoareyou.rs
@@ -37,9 +37,9 @@ impl WhoAreYou {
 
         let mut frame = Frame::array();
         frame.push_bulk(Bytes::from(msg_type.as_bytes()));
-        frame.push_int(self.src_p2p_port as u64);
+        frame.push_int(self.src_p2p_port as u128);
         frame.push_bulk(src_sig_bytes.into());
-        frame.push_int(self.src_disc_port as u64);
+        frame.push_int(self.src_disc_port as u128);
         frame.push_bulk(src_public_key_bytes.into());
 
         Ok(frame)
